requirements.txt
asgiref==3.11.1
dj-rest-auth==7.2.0
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
